+++ web/src/main/java/io/studytracker/config/properties/GitLabProperties.java
@@ -31,7 +31,7 @@ public class GitLabProperties {
 
   private URL url;
 
-  private String accessKey;
+  private String accessToken;
 
   private Integer rootGroupId;
 
@@ -49,7 +49,7 @@ public class GitLabProperties {
   public String toString() {
     return "GitLabProperties{" +
         "url=" + url +
-        ", accessKey='" + accessKey.substring(0,5) + "*****'" +
+        ", accessKey='" + accessToken.substring(0,5) + "*****'" +
         ", rootGroupId=" + rootGroupId +
         ", clientId='" + clientId + '\'' +
         ", clientSecret='" + clientSecret + '\'' +