+++ CHANGELOG.md
@@ -22,9 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Also overhauled how storage folder records are managed on the back-end, allowing for more flexible folder management across multiple file storage services.
 - Consolidated assay and study controls into a single dropdown in the Actions widget.
 - Study and assay names can now be changed by editing them. This name change will *not* cascade to storage or notebook folders created with the original names.
+- Updated the styling of the program list and details pages.
 
 ### Fixed
 - Fixed several issues with File Manager breadcrumbs and controls, which prevented them from acting as expected.
+- Fixed an issue where large ELN folder requests could timeout or crash study & assay creation requests.
 
 ## [0.8.2] - 2023-03-01